Updated to VS 2022 and removed compilation warnings

+++ TUIO/TuioBlob.cpp
@@ -107,20 +107,20 @@ void TuioBlob::update (TuioTime ttime, float xp, float yp, float a, float w, flo
 	float last_rotation_speed = rotation_speed;
 	
 	float da = a-angle;
-	if (da > M_PI) da -= 2*M_PI;
-	else if (da < -M_PI) da+=2*M_PI;
+	if (da > M_PI) da -= 2* (float)M_PI;
+	else if (da < -M_PI) da+=2*(float)M_PI;
 	
 	float prev_angle = angle;
 	if (angleFilter) angle = angleFilter->filter(angle+da,dt);
 	else angle = angle+da;
 	if (fabs(angle-prev_angle)<angleThreshold) angle = prev_angle;
 	
-	if (angle > 2*M_PI) angle-=2*M_PI;
-	else if (angle < 0) angle+=2*M_PI;
+	if (angle > 2*M_PI) angle-=2* (float)M_PI;
+	else if (angle < 0) angle+=2* (float)M_PI;
 	
 	da = angle-prev_angle;
-	if (da > M_PI) da -= 2*M_PI;
-	else if (da < -M_PI) da+=2*M_PI;
+	if (da > M_PI) da -= 2* (float)M_PI;
+	else if (da < -M_PI) da+=2* (float)M_PI;
 	da = da/(2*M_PI);
 	
 	if (widthFilter && heightFilter) {
